polepad-ai/apps/ocr-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pytesseract==0.3.10
pillow==10.3.0
numpy==1.26.4
pydantic==2.7.1
httpx==0.27.0

